move LRUdict in utils.py
